moved contact specs into a folder
